Enhancements to check-apt-upgrade.pl
Added support for generating Nagios Performance Data
Added command line arguments to support setting warning and critical values for the Nagios Performance data feature. 
The warning and critical values can be set separately for the upgrades and the security updates. 
Added support to allow the Nagios status of OK in the event that number of updates is below the warning value. 
Changed numbered captures to named captures.